Store Fiber.current in a property
By getting rid of the Fiber.current getter and just resetting it every
time it changes we can drastically speed up access to that property,
since it saves a trip out of the v8 engine. In most cases it seems like
people will be using Fiber.current, so this is a good change to make.