More pthread hook fixing in Lion
The approach from the last commit wasn't totally good enough. It seems
v8 implements "fast TLS" in some cases which uses inline assembly to
get at thread locals and there's no way we can hook this. But this is
only being used for Isolates so they don't need to be fiber-specific.
To solve the problem when setting a local in FLS it also sets the local
in FLS. ** This means that in fibers GetExistingThreadLocal and
GetThreadLocal will have conflicting results ** but so far I haven't
had any problems.